feat(contracts): coagulation panel (PT, INR, aPTT, Quick)
Closes #352. Routine ER bloodwork CBC bundle's coagulation half: PT,
INR, aPTT, aPTT ratio, and the DACH/ES/FR Quick index. Anticoagulant
monitoring (warfarin via INR, heparin via aPTT) and bleeding-risk screen
were previously unsupported — surfaced by the Sagrat Cor discharge audit.

Mechanics:
- Five new MetricType keys (PT/INR/Quick/aPTT/aPTT-ratio); INR and
  aPTT-ratio use SCORE per the HOMA_IR dimensionless-index precedent.
- LOINC mappings (5902-2 / 6301-6 / 11149-6 / 14979-9 / 13488-2).
- INR gets bidirectional bands (ACCP 2012 universal alarms: <0.8 low,
  >=5.0 bleed-risk high). PT/aPTT/Quick stay reference-range-descriptive —
  therapeutic windows depend on anticoagulant indication and belong to
  the owner and their clinician.
- New "Coagulation" BiomarkerPanel row on the dashboard, ordered after
  hematology where lab reports group them.

Side cleanup: extracted loincCode/ucumCode tables from FhirExporter.kt
(was at 500 lines, would have overflowed with the additions) to a new
FhirMetricCodings.kt. The FhirExporterTest mirror functions are dropped
in favour of calling the production internal functions directly — the
mirror was a workaround for the same length pressure that no longer
applies. Net diff: -228 lines.
